0.6.8 (vc29): HLC counter Int->Long + session handoff
Counter exhaustion under a wall pin (user probe): every node increments
on every event while pinned, crossing an Int-scale sanity ceiling in
weeks, after which peers reject legit stamps and causality breaks. Long
counter + 1e15 ceiling makes exhaustion unreachable. Wire is fleet-only
so the format change is free; old parsers null out oversized counters
harmlessly. Fleet flashed together.

Handoff: session wrap prepended (echo loop dead, _ext column fix, O124,
dead-group re-bootstrap, O114->G41, O95->G42, standing directives).

By Order Of The High Magnate

+++ app/src/main/java/com/rumor/mesh/service/MeshService.kt
@@ -263,11 +263,11 @@ class MeshService : Service(), MeshController {
         gossipEngine.hlc.restore(
             com.rumor.mesh.core.time.HlcTimestamp(
                 hlcPrefs.getLong("wallMs", 0L),
-                hlcPrefs.getInt("counter", 0),
+                hlcPrefs.getLong("counter", 0L),
             )
         )
         gossipEngine.hlc.onAdvance = { ts ->
-            hlcPrefs.edit().putLong("wallMs", ts.wallMs).putInt("counter", ts.counter).apply()
+            hlcPrefs.edit().putLong("wallMs", ts.wallMs).putLong("counter", ts.counter).apply()
         }
 
         // O124: answer a pulse from an unknown-or-stale peer with our own
